refactor(agent): extract package broker into dedicated crate
Move the package broker subsystem out of the devolutions-agent binary
into a new crates/now-package-broker crate, keeping the agent manifest
lean and giving the broker-only dependency set (axum, hyper-util,
notify, now-policy, ...) its own home. The crate compiles to an empty
library on non-Windows platforms.

Extract code_signing into devolutions-agent-shared so it is shared
between the updater and the broker, which previously reached back into
the agent crate for it. The agent forwards the development-only
dev-skip-broker-signature feature to the broker crate.

Issue: DGW-417
