+++ crates/now-package-broker/src/assets/samples/deny-risky-options.policy.json
