+++ crates/now-package-broker/src/assets/samples/requests/winget-unknown-install.request.json
