+++ crates/now-package-broker/src/assets/samples/requests/winget-vscode-install.request.json
