+++ crates/now-package-broker/src/assets/samples/requests/winget-vscode-skiphash.request.json
