+++ crates/now-package-broker/src/assets/samples/scenarios/baseline.scenarios.json
