+++ crates/now-package-broker/src/command_builder/dotnet.rs
