+++ crates/now-package-broker/src/executor/windows/privileges.rs
