+++ devolutions-agent/Cargo.toml
@@ -13,7 +13,7 @@ default = []
 # Development-only feature allowing the SkipBrokerSignatureValidation debug option to take effect.
 # Must never be enabled for shipped builds: without it, broker client signature validation is
 # unconditionally enforced regardless of the configuration file contents.
-dev-skip-broker-signature = []
+dev-skip-broker-signature = ["now-package-broker/dev-skip-broker-signature"]
 
 [lints]
 workspace = true
@@ -23,12 +23,10 @@ agent-tunnel-proto = { path = "../crates/agent-tunnel-proto" }
 anyhow = "1"
 backoff = "0.4"
 async-trait = "0.1"
-axum = { version = "0.8", default-features = false, features = ["http1", "json", "tokio", "original-uri", "matched-path"] }
 base64 = "0.23"
 bytes = "1"
 camino = { version = "1.1", features = ["serde1"] }
 ceviche = "0.7"
-chrono = { version = "0.4", features = ["serde"] }
 ctrlc = "3.5"
 devolutions-agent-shared = { path = "../crates/devolutions-agent-shared" }
 devolutions-gateway-task = { path = "../crates/devolutions-gateway-task" }
@@ -36,21 +34,14 @@ devolutions-log = { path = "../crates/devolutions-log" }
 futures = "0.3"
 hex = "0.4"
 hostname = "0.4"
-hyper = { version = "1", features = ["http1", "server"] }
-hyper-util = { version = "0.1", features = ["tokio", "server", "server-auto", "service"] }
 http-client-proxy = { path = "../crates/http-client-proxy" }
 ipnetwork = "0.20"
-notify = { version = "7", default-features = false, features = ["macos_kqueue"] }
-now-policy = "0.2"
-now-policy-api = { version = "0.3", features = ["policy-compat"] }
-now-policy-server-template = { version = "0.3", features = ["policy-compat"] }
 parking_lot = "0.12"
 prost = "0.13"
 prost-types = "0.13"
 quinn = "0.11"
 rand = "0.8" # FIXME(@CBenoit): maybe we don't need this crate
 rcgen = { version = "0.13", features = ["pem"] }
-regex = "1"
 reqwest = { version = "0.12", default-features = false, features = ["rustls-tls-native-roots", "http2", "socks", "json"] }
 rustls = { version = "0.23", default-features = false, features = ["std", "ring"] }
 rustls-pemfile = "2.2"
@@ -58,15 +49,11 @@ rustls-pki-types = "1"
 sha2 = "0.10"
 serde_json = "1"
 serde = { version = "1", features = ["derive"] }
-serde_yaml = "0.9"
-semver = "1"
 tap = "1.0"
 tempfile = "3"
 tokio-rustls = { version = "0.26", default-features = false, features = ["logging", "tls12", "ring"] }
 tokio-stream = "0.1"
-tokio-util = "0.7"
 tonic = { version = "0.12", features = ["transport"] }
-tower-service = "0.3"
 tracing = "0.1"
 url = { version = "2.5", features = ["serde"] }
 x509-parser = "0.16"
@@ -101,10 +88,10 @@ aws-lc-rs = "1.15"
 time = { version = "0.3", features = ["local-offset", "macros", "parsing"] }
 devolutions-pedm = { path = "../crates/devolutions-pedm" }
 notify-debouncer-mini = "0.6"
+now-package-broker = { path = "../crates/now-package-broker" }
 reqwest = { version = "0.12", default-features = false, features = ["rustls-tls-native-roots", "http2", "socks"] }
 thiserror = "2"
 uuid = { version = "1.17", features = ["v4"] }
-widestring = "1.2"
 win-api-wrappers = { path = "../crates/win-api-wrappers" }
 
 [target.'cfg(windows)'.dependencies.windows]
@@ -118,13 +105,7 @@ features = [
     "Win32_Security_Cryptography",
     "Win32_Security_Authorization",
     "Win32_System_ApplicationInstallationAndServicing",
-    "Win32_System_Ioctl",
-    "Win32_System_IO",
-    "Win32_System_Pipes",
     "Win32_System_RemoteDesktop",
-    "Win32_System_Com",
-    "Win32_System_Console",
-    "Win32_UI_Shell",
 ]
 
 [target.'cfg(windows)'.build-dependencies]
