--- devolutions-agent/src/broker/mod.rs
@@ -1,20 +1,0 @@
-//! Package broker module.
-//!
-//! Provides policy evaluation and command execution for package operations,
-//! communicating over a Windows named pipe using HTTP/1.1.
-
-pub(crate) mod auth;
-pub mod command_builder;
-pub mod evaluator;
-pub mod event_channel;
-pub mod executor;
-pub mod operation_tracker;
-pub mod pipe;
-pub mod policy_loader;
-pub(crate) mod policy_security;
-pub mod policy_watcher;
-pub mod server;
-pub mod task;
-
-#[cfg(test)]
-mod scenario_tests;
